feat: Add hover effect to color items and font style
Changes were made in CSS to maintain Arial as the default font, as well as to enhance the design of the color-item class. These changes included adding a hover effect, displaying the color RGB value in a hidden paragraph which is visible on mouseover, employing a flex layout, and more.

For `index.html`, the prior individual div-elements representing colors were replaced with JS to improve maintainability and scalability. Now colors can be added/removed easily by manipulating the colors array.